join League, create league, leave league

+++ src/Common/CleanArchitecture.Application/Leagues/Commands/LeaveLeague/LeaveLeagueCommand.cs
@@ -30,8 +30,11 @@ public class LeaveLeagueCommandHandler : IRequestHandlerWrapper<LeaveLeagueComma
         var userProfile = await _context.UserProfiles.FirstOrDefaultAsync(x=>x.UserId == _currentUserService.UserId, cancellationToken);
         
         var membership = await _context.LeagueMemberships.Where(x=>x.LeagueId == request.LeagueId && x.UserProfileId == userProfile.Id).FirstOrDefaultAsync(cancellationToken);
-        _context.LeagueMemberships.Remove(membership);
-        await _context.SaveChangesAsync(cancellationToken);
+        if (membership is not null)
+        {
+            _context.LeagueMemberships.Remove(membership);
+            await _context.SaveChangesAsync(cancellationToken);    
+        }
         return ServiceResult.Success(true);
     }
 }
